feat: add utils to download the manifest
